refactor: harden usage drain report readiness
