+++ MANIFEST.in
@@ -2,6 +2,8 @@ include README.md LICENSE CHANGELOG.md CONTRIBUTING.md SECURITY.md AGENTS.md .mc
 recursive-include .codex-plugin *.json
 recursive-include assets *.svg
 recursive-include docs *.md *.png
+exclude docs/usage-drain-research-white-paper.md
+prune output
 recursive-include skills *.md *.py
 recursive-include scripts *.py
 recursive-include src/codex_usage_tracker/plugin_data/assets *
