+++ pyproject.toml
@@ -104,4 +104,6 @@ files = [
   "src/codex_usage_tracker/reports.py",
   "src/codex_usage_tracker/schema.py",
   "src/codex_usage_tracker/store.py",
+  "src/codex_usage_tracker/usage_drain_reports.py",
+  "src/codex_usage_tracker/usage_drain_time_series.py",
 ]
